erasmus + partner embed & alle embeds samengevoegd in images.go
